Made separate files for util functions, to prevent excessive code duplication in migrations
